passing vector from int to float [UNSTABLE]

+++ srcs/structures/engine/map_params/utils/parsing/map/is_walled.c
@@ -16,9 +16,9 @@ static t_bool   is_up_walled(char ** map, t_vector index)
 {
     while (index.y >= 0)
     {
-        if (is_tile(map[index.y][index.x]) == false)
+        if (is_tile(map[(int)index.y][(int)index.x]) == false)
         {
-            if (map[index.y][index.x] == '1')
+            if (map[(int)index.y][(int)index.x] == '1')
                 return (true);
             else
                 return (false);
@@ -32,9 +32,9 @@ static t_bool   is_down_walled(char ** map, t_vector index)
 {
     while (index.y < ft_tab_len((void **)map))
     {
-        if (is_tile(map[index.y][index.x]) == false)
+        if (is_tile(map[(int)index.y][(int)index.x]) == false)
         {
-            if (map[index.y][index.x] == '1')
+            if (map[(int)index.y][(int)index.x] == '1')
                 return (true);
             else
                 return (false);
@@ -46,11 +46,11 @@ static t_bool   is_down_walled(char ** map, t_vector index)
 
 static t_bool   is_right_walled(char **map, t_vector index)
 {
-    while (index.x < (int)(ft_strlen(map[index.y])))
+    while (index.x < (int)(ft_strlen(map[(int)index.y])))
     {
-        if (is_tile(map[index.y][index.x]) == false)
+        if (is_tile(map[(int)index.y][(int)index.x]) == false)
         {
-            if (map[index.y][index.x] == '1')
+            if (map[(int)index.y][(int)index.x] == '1')
                 return (true);
             else
                 return (false);
@@ -64,9 +64,9 @@ static t_bool   is_left_walled(char **map, t_vector index)
 {
     while (index.x >= 0)
     {
-        if (is_tile(map[index.y][index.x]) == false)
+        if (is_tile(map[(int)index.y][(int)index.x]) == false)
         {
-            if (map[index.y][index.x] == '1')
+            if (map[(int)index.y][(int)index.x] == '1')
                 return (true);
             else
                 return (false);
